Fix partial rendering position with fluid slot

+++ src/main/java/com/cleanroommc/modularui/widgets/FluidSlot.java
@@ -128,11 +128,11 @@ public class FluidSlot extends Widget<FluidSlot> implements Interactable, JeiGho
         IFluidTank fluidTank = getFluidTank();
         FluidStack content = this.syncHandler.getValue();
         if (content != null) {
-            int y = this.contentOffsetY;
+            float y = this.contentOffsetY;
             float height = getArea().height - y * 2;
             if (!this.alwaysShowFull) {
                 float newHeight = height * content.amount * 1f / fluidTank.getCapacity();
-                y += (int) (height - newHeight);
+                y += height - newHeight;
                 height = newHeight;
             }
             GuiDraw.drawFluidTexture(content, this.contentOffsetX, y, getArea().width - this.contentOffsetX * 2, height, 0);
